+++ applications/fishing-map/src/features/timebar/TimebarSettings.tsx
@@ -41,7 +41,9 @@ const TimebarSettings = () => {
     dispatchQueryParams({ timebarEvents: o.id as TimebarEvents })
   }
   const setGraphOption = (o: SelectOption) => {
-    dispatchQueryParams({ timebarGraph: o.id as TimebarGraphs })
+    if (!o.label.includes('Coming soon')) {
+      dispatchQueryParams({ timebarGraph: o.id as TimebarGraphs })
+    }
   }
   const removeEventsOption = () => {
     dispatchQueryParams({ timebarEvents: TimebarEvents.None })
@@ -50,7 +52,6 @@ const TimebarSettings = () => {
     dispatchQueryParams({ timebarGraph: TimebarGraphs.None })
   }
   const expandedContainerRef = useClickedOutside(closeOptions)
-
   if (!timebarVisualisation) return null
   return (
     <div className={styles.container} ref={expandedContainerRef}>
@@ -67,46 +68,30 @@ const TimebarSettings = () => {
       {optionsPanelOpen && (
         <div className={styles.optionsContainer}>
           {activeHeatmapDataviews && activeHeatmapDataviews.length > 0 && (
-            <div>
-              <Radio
-                active={timebarVisualisation === TimebarVisualisations.Heatmap}
-                onClick={setHeatmapActive}
-              />
-              <span
-                className={cx(styles.option, {
-                  [styles.active]: timebarVisualisation === TimebarVisualisations.Heatmap,
-                })}
-                onClick={setHeatmapActive}
-              >
-                Apparent Fishing Effort
-              </span>
-            </div>
+            <Radio
+              label={t('common.apparentFishing', 'Apparent Fishing Effort')}
+              active={timebarVisualisation === TimebarVisualisations.Heatmap}
+              onClick={setHeatmapActive}
+            />
           )}
           {activeVesselDataviews && activeVesselDataviews.length > 0 && (
             <Fragment>
-              <div>
-                <Radio
-                  active={timebarVisualisation === TimebarVisualisations.Vessel}
-                  onClick={setVesselActive}
-                />
-                <span
-                  className={cx(styles.option, {
-                    [styles.active]: timebarVisualisation === TimebarVisualisations.Vessel,
-                  })}
-                  onClick={setVesselActive}
-                >
-                  {t('vessel.tracks', 'Vessel Tracks')}
-                </span>
-              </div>
+              <Radio
+                label={t('vessel.tracks', 'Vessel Tracks')}
+                active={timebarVisualisation === TimebarVisualisations.Vessel}
+                onClick={setVesselActive}
+              />
               {timebarVisualisation === TimebarVisualisations.Vessel && (
                 <div className={styles.vesselTrackOptions}>
                   <Select
-                    label={t('common.events', 'Events')}
+                    // label={t('common.events', 'Events')}
+                    label="Events (Coming soon)"
                     options={TIMEBAR_EVENT_OPTIONS}
                     selectedOption={TIMEBAR_EVENT_OPTIONS.find((o) => o.id === timebarEvents)}
                     onSelect={setEventsOption}
                     onRemove={removeEventsOption}
                     direction="top"
+                    disabled
                   />
                   <Select
                     label={t('timebar.graph', 'Graph')}