+++ packages/ui-components/src/select/Select.tsx
@@ -17,6 +17,7 @@ interface SelectProps {
   onCleanClick?: (e: React.MouseEvent) => void
   className?: string
   direction?: 'bottom' | 'top'
+  disabled?: boolean
 }
 
 const isItemSelected = (selectedItem: SelectOption | undefined, item: SelectOption) => {
@@ -34,6 +35,7 @@ function Select(props: SelectProps) {
     onCleanClick,
     className,
     direction = 'bottom',
+    disabled = false,
   } = props
   const {
     isOpen,
@@ -46,7 +48,7 @@ function Select(props: SelectProps) {
   } = useSelect<SelectOption | null>({
     items: options,
     onSelectedItemChange: ({ selectedItem }) => {
-      if (selectedItem) {
+      if (!disabled && selectedItem) {
         handleChange(selectedItem)
         selectItem(null)
       }
@@ -63,7 +65,9 @@ function Select(props: SelectProps) {
     },
     [onRemove, onSelect, selectedOption]
   )
+
   const hasSelectedOptions = selectedOption !== undefined
+
   return (
     <div>
       <label {...getLabelProps()}>{label}</label>
@@ -94,17 +98,19 @@ function Select(props: SelectProps) {
             options.map((item, index) => {
               const highlight = highlightedIndex === index
               const selected = isItemSelected(selectedOption, item)
+              const itemDisabled = disabled || item.disabled
               return (
                 <Tooltip key={`${item}${index}`} content={item.tooltip} placement="top-start">
                   <li
                     className={cx(styles.optionItem, {
                       [styles.selected]: selected,
                       [styles.highlight]: highlight,
+                      [styles.disabled]: itemDisabled,
                     })}
                     {...getItemProps({ item, index })}
                   >
                     {item.label}
-                    {highlight && <Icon icon={selected ? 'close' : 'tick'} />}
+                    {highlight && !itemDisabled && <Icon icon={selected ? 'close' : 'tick'} />}
                   </li>
                 </Tooltip>
               )